requirements.txt
# API Hunter - Advanced Bug Bounty Tool Dependencies

# HTTP & Networking
requests>=2.32.3
httpx>=0.25.0
aiohttp>=3.9.0
websockets>=12.0
grpcio>=1.59.0

# CLI & UI Framework
click>=8.1.7
rich>=14.0.0
textual>=0.45.0
typer>=0.9.0

# Data Processing & Validation
pydantic>=2.5.0
jsonschema>=4.20.0
python-multipart>=0.0.6
lxml>=4.9.3
beautifulsoup4>=4.12.2

# Database & Caching
sqlalchemy>=2.0.23
alembic>=1.13.0
redis>=5.0.1
psycopg2-binary>=2.9.8

# Security & Crypto
cryptography>=41.0.7
pyjwt>=2.8.0
passlib>=1.7.4
python-jose>=3.3.0

# AI & ML
openai>=1.3.0
transformers>=4.35.0
sentence-transformers>=2.2.2
chromadb>=0.4.0

# Task Queue & Background Jobs
celery>=5.3.4
flower>=2.0.1
kombu>=5.3.4

# Testing & Quality
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Proxy & Interception
mitmproxy>=10.1.5

# Utilities
python-dotenv>=1.0.0
jinja2>=3.1.2
weasyprint>=60.2
pyyaml>=6.0.1
uuid>=1.30